+++ lib/notifications/client/response_precompiled_letter.rb
@@ -1,0 +1,19 @@
+module Notifications
+  class Client
+    class ResponsePrecompiledLetter
+      FIELDS = %i(
+      id
+      reference
+      postage
+    ).freeze
+
+      attr_reader(*FIELDS)
+
+      def initialize(notification)
+        FIELDS.each do |field|
+          instance_variable_set(:"@#{field}", notification.fetch(field.to_s, nil))
+        end
+      end
+    end
+  end
+end